requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-environ==0.4.5
environ==1.0
idna==2.10
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
spoonacular==1.0.0
sqlparse==0.3.1
urllib3==1.25.9
